update an error in si_dos.in
Hi, thanks so much for the tutorials. Just to correct a bug in the si_dos.in file: 

There should be an `EOL` at the end to indicate the end of the namelist, otherwise it prompts error in "reading namelist". The strange thing was orinally spotted in 2018 (https://lists.quantum-espresso.org/pipermail/users/2018-December/041756.html), and it still occurs in qe-7.1 version. So I make that change, please accept the correction if you think it would be helpful. Thanks!